add new color (cyan) to the list of known colors when the per contact send mode is set to ForceAnsi, tabSRMM will write Proto/UnicodeSend and set it to 0 before sending a message (works with ICQJ to solve problems with trillian)
git-svn-id: http://miranda.googlecode.com/svn/trunk@2664 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/msgdialog.c
@@ -2315,7 +2315,7 @@ BOOL CALLBACK DlgProcMessage(HWND hwndDlg, UINT msg, WPARAM wParam, LPARAM lPara
 
 				GetWindowRect(hwndDlg, &rcWin);
 				GetClientRect(hwndDlg, &rcClient);
-				newMessagePos = rcWin.bottom - (short)wParam;
+				newMessagePos = (short)rcWin.bottom - (short)wParam;
 
 
 				SendMessage(hwndDlg, DM_SAVESIZE, 0, 0);
